Replace 'unsigned', 'int' and 'unsigned int' with 'size_t'
.. where appropriate.

+++ hlslang/GLSLCodeGen/glslSymbol.cpp
@@ -200,7 +200,7 @@ static bool IsReservedGlslKeyword (const std::string& name)
 }
 
 
-GlslSymbol::GlslSymbol( const std::string &n, const std::string &s, int id, EGlslSymbolType t, TPrecision prec, EGlslQualifier q, int as ) : GlslSymbolOrStructMemberBase(n, s, t, q, prec, as),
+GlslSymbol::GlslSymbol( const std::string &n, const std::string &s, int id, EGlslSymbolType t, TPrecision prec, EGlslQualifier q, size_t as ) : GlslSymbolOrStructMemberBase(n, s, t, q, prec, as),
    identifier(id),
    mangleCounter(0),
    structPtr(0),
